requirements.txt
asn1crypto==0.24.0
Authlib==0.10
autopep8==1.4.3
Babel==2.4.0
boto3==1.9.124
botocore==1.12.124
braintree==3.52.0
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
Click==7.0
colorama==0.4.1
cryptography==2.6.1
cssmin==0.2.0
docutils==0.14
Flask==1.0.2
Flask-Assets==0.12
Flask-Cors==3.0.7
Flask-JWT==0.3.2
Flask-Login==0.4.1
Flask-WTF==0.14.2
gunicorn==19.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
jmespath==0.9.4
jsmin==2.2.2
MarkupSafe==1.1.0
peewee==3.1.0
peewee-db-evolve==3.7.0
peewee-validates==1.0.7
psycopg2-binary==2.7.7
pycodestyle==2.5.0
pycparser==2.19
PyJWT==1.4.2
python-dateutil==2.8.0
python-dotenv==0.10.1
python-http-client==3.1.0
pytz==2018.9
redis==3.2.1
requests==2.21.0
rq==0.13.0
s3transfer==0.2.0
sendgrid==5.6.0
six==1.12.0
urllib3==1.24.1
urlparse3==1.1
webassets==0.12.1
Werkzeug==0.14.1
WTForms==2.2.1
